+++ js/pkg/garu_wasm_bg.wasm.d.ts
@@ -4,7 +4,7 @@ export const memory: WebAssembly.Memory;
 export const __wbg_garuwasm_free: (a: number, b: number) => void;
 export const garuwasm_analyze: (a: number, b: number, c: number, d: number) => void;
 export const garuwasm_analyze_topn: (a: number, b: number, c: number, d: number, e: number) => void;
-export const garuwasm_new: (a: number, b: number, c: number, d: number, e: number) => void;
+export const garuwasm_new: (a: number, b: number, c: number) => void;
 export const garuwasm_tokenize: (a: number, b: number, c: number, d: number) => void;
 export const garuwasm_version: (a: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;